+++ canopen_sniffer/dox/documentation/html/namespacecanopen__sniffer.js
@@ -2,6 +2,8 @@ var namespacecanopen__sniffer =
 [
     [ "bus_stats", "classcanopen__sniffer_1_1bus__stats.html", "classcanopen__sniffer_1_1bus__stats" ],
     [ "can_sniffer", "classcanopen__sniffer_1_1can__sniffer.html", "classcanopen__sniffer_1_1can__sniffer" ],
+    [ "DisplayCLI", "classcanopen__sniffer_1_1DisplayCLI.html", "classcanopen__sniffer_1_1DisplayCLI" ],
+    [ "DisplayGUI", "classcanopen__sniffer_1_1DisplayGUI.html", "classcanopen__sniffer_1_1DisplayGUI" ],
     [ "eds_parser", "classcanopen__sniffer_1_1eds__parser.html", "classcanopen__sniffer_1_1eds__parser" ],
     [ "frame_type", "classcanopen__sniffer_1_1frame__type.html", null ],
     [ "process_frame", "classcanopen__sniffer_1_1process__frame.html", "classcanopen__sniffer_1_1process__frame" ],
@@ -15,10 +17,12 @@ var namespacecanopen__sniffer =
     [ "CLI_DATA_TABLE_HEIGHT", "namespacecanopen__sniffer.html#a7efe84abab5551614e0621ca15cee0ec", null ],
     [ "CLI_GRAPH_WIDTH", "namespacecanopen__sniffer.html#aa46a1dd1fb2dd11e460a3472c7c6b7ad", null ],
     [ "CLI_PROTOCOL_TABLE_HEIGHT", "namespacecanopen__sniffer.html#adb2fef007c95e5c7bad348938406dc9b", null ],
-    [ "DEFAULT_BAUD_RATE", "namespacecanopen__sniffer.html#a103859a1dddc0923cb7ced9ecafffe94", null ],
+    [ "DEFAULT_CAN_BIT_RATE", "namespacecanopen__sniffer.html#a1c22777dfecff81d656d3c48e0a18187", null ],
     [ "DEFAULT_INTERFACE", "namespacecanopen__sniffer.html#a9313201603fcc86a7436c761ac54a0af", null ],
     [ "FILENAME", "namespacecanopen__sniffer.html#ae4e6f99e1b73b1f62bb9ed8ced355214", null ],
     [ "FSYNC_EVERY", "namespacecanopen__sniffer.html#abd5e97aaf326e2debb2e68277808c957", null ],
     [ "log", "namespacecanopen__sniffer.html#ab09a63eb35b270b5cdbead1983ebdccb", null ],
-    [ "MAX_FRAMES", "namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc", null ]
+    [ "LOG_LEVEL", "namespacecanopen__sniffer.html#aa5a9053636a30269210c54e734e0d583", null ],
+    [ "MAX_FRAMES", "namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc", null ],
+    [ "root", "namespacecanopen__sniffer.html#ab4b8daf4b8ea9d39568719e1e320076f", null ]
 ];