+++ canopen_sniffer/dox/documentation/html/search/all_7.js
@@ -1,6 +1,7 @@
 var searchData=
 [
-  ['get_5fframe_5fcount_39',['get_frame_count',['../classcanopen__sniffer_1_1bus__stats.html#a04a08b7d5a675eaddf9ac350966183c4',1,'canopen_sniffer::bus_stats']]],
-  ['get_5fsnapshot_40',['get_snapshot',['../classcanopen__sniffer_1_1bus__stats.html#aec57c1b7a69ec3efeb622f269d4484f0',1,'canopen_sniffer::bus_stats']]],
-  ['get_5ftotal_5fframes_41',['get_total_frames',['../classcanopen__sniffer_1_1bus__stats.html#a4204bbed3351741f6e167302b564c36c',1,'canopen_sniffer::bus_stats']]]
+  ['get_5fframe_5fcount_49',['get_frame_count',['../classcanopen__sniffer_1_1bus__stats.html#a04a08b7d5a675eaddf9ac350966183c4',1,'canopen_sniffer::bus_stats']]],
+  ['get_5fsnapshot_50',['get_snapshot',['../classcanopen__sniffer_1_1bus__stats.html#aec57c1b7a69ec3efeb622f269d4484f0',1,'canopen_sniffer::bus_stats']]],
+  ['get_5ftotal_5fframes_51',['get_total_frames',['../classcanopen__sniffer_1_1bus__stats.html#a4204bbed3351741f6e167302b564c36c',1,'canopen_sniffer::bus_stats']]],
+  ['graph_5fwidth_52',['GRAPH_WIDTH',['../classcanopen__sniffer_1_1DisplayCLI.html#a3f49a3f3cb41debc86ca3d71a7ad500c',1,'canopen_sniffer::DisplayCLI']]]
 ];