+++ canopen_sniffer/dox/documentation/html/search/all_b.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['main_52',['main',['../namespacecanopen__sniffer.html#a355f06d5ba7ea264dc3e2e128349c130',1,'canopen_sniffer']]],
-  ['max_5fframes_53',['MAX_FRAMES',['../namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'canopen_sniffer']]]
+  ['main_65',['main',['../namespacecanopen__sniffer.html#a355f06d5ba7ea264dc3e2e128349c130',1,'canopen_sniffer']]],
+  ['max_5fframes_66',['MAX_FRAMES',['../namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'canopen_sniffer']]]
 ];