+++ canopen_sniffer/dox/documentation/html/search/all_e.js
@@ -1,7 +1,10 @@
 var searchData=
 [
-  ['raw_5fframe_63',['raw_frame',['../classcanopen__sniffer_1_1can__sniffer.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'canopen_sniffer.can_sniffer.raw_frame()'],['../classcanopen__sniffer_1_1process__frame.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'canopen_sniffer.process_frame.raw_frame()']]],
-  ['readme_2emd_64',['README.md',['../README_8md.html',1,'']]],
-  ['reset_65',['reset',['../classcanopen__sniffer_1_1bus__stats.html#a51829b63adb24ac48d350dee60181002',1,'canopen_sniffer::bus_stats']]],
-  ['run_66',['run',['../classcanopen__sniffer_1_1can__sniffer.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.can_sniffer.run()'],['../classcanopen__sniffer_1_1process__frame.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.process_frame.run()']]]
+  ['rates_5fstats_80',['rates_stats',['../classcanopen__sniffer_1_1bus__stats_1_1rates__stats.html',1,'canopen_sniffer::bus_stats']]],
+  ['raw_5fframe_81',['raw_frame',['../classcanopen__sniffer_1_1can__sniffer.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'canopen_sniffer.can_sniffer.raw_frame()'],['../classcanopen__sniffer_1_1process__frame.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'canopen_sniffer.process_frame.raw_frame()']]],
+  ['readme_2emd_82',['README.md',['../README_8md.html',1,'']]],
+  ['render_5ftables_83',['render_tables',['../classcanopen__sniffer_1_1DisplayCLI.html#a1602e5a067d7cbe99e22f10101f9e381',1,'canopen_sniffer::DisplayCLI']]],
+  ['reset_84',['reset',['../classcanopen__sniffer_1_1bus__stats.html#a51829b63adb24ac48d350dee60181002',1,'canopen_sniffer::bus_stats']]],
+  ['root_85',['root',['../namespacecanopen__sniffer.html#ab4b8daf4b8ea9d39568719e1e320076f',1,'canopen_sniffer']]],
+  ['run_86',['run',['../classcanopen__sniffer_1_1can__sniffer.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.can_sniffer.run()'],['../classcanopen__sniffer_1_1process__frame.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.process_frame.run()'],['../classcanopen__sniffer_1_1DisplayCLI.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.DisplayCLI.run()'],['../classcanopen__sniffer_1_1DisplayGUI.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.DisplayGUI.run()']]]
 ];