+++ canopen_sniffer/dox/documentation/html/search/all_f.js
@@ -1,15 +1,17 @@
 var searchData=
 [
-  ['save_5fframe_67',['save_frame',['../classcanopen__sniffer_1_1process__frame.html#a330d0aa4e889583805485fc069b8d7e3',1,'canopen_sniffer::process_frame']]],
-  ['save_5fframe_5fto_5fcsv_68',['save_frame_to_csv',['../classcanopen__sniffer_1_1can__sniffer.html#adea2a4942ab2e9b485f0f4acb670a172',1,'canopen_sniffer.can_sniffer.save_frame_to_csv()'],['../classcanopen__sniffer_1_1process__frame.html#a43c1277547b2e3a9db6631c8134c28f6',1,'canopen_sniffer.process_frame.save_frame_to_csv(self, int cob, frame_type ftype, int index, int sub, str name, str raw, str decoded)']]],
-  ['save_5fprocessed_5fframe_69',['save_processed_frame',['../classcanopen__sniffer_1_1process__frame.html#a8a398b1c620b611b45d8da94d9d35e6c',1,'canopen_sniffer::process_frame']]],
-  ['sdo_5freq_70',['SDO_REQ',['../classcanopen__sniffer_1_1frame__type.html#aa227b2666e9ee4b630add48d6dec0c39',1,'canopen_sniffer::frame_type']]],
-  ['sdo_5fres_71',['SDO_RES',['../classcanopen__sniffer_1_1frame__type.html#a95373d5028d5cb3c1ce7b6120de91447',1,'canopen_sniffer::frame_type']]],
-  ['sdo_5fstats_72',['sdo_stats',['../classcanopen__sniffer_1_1bus__stats_1_1sdo__stats.html',1,'canopen_sniffer::bus_stats']]],
-  ['set_73',['set',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html#a112a711f766db446c2f517b794a1c04a',1,'canopen_sniffer::bus_stats::stats_data']]],
-  ['set_5fstart_5ftime_74',['set_start_time',['../classcanopen__sniffer_1_1bus__stats.html#a09ae5bd534d8dbec99a913ed75fe9848',1,'canopen_sniffer::bus_stats']]],
-  ['stats_75',['stats',['../classcanopen__sniffer_1_1process__frame.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer::process_frame']]],
-  ['stats_5fdata_76',['stats_data',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html',1,'canopen_sniffer::bus_stats']]],
-  ['stop_77',['stop',['../classcanopen__sniffer_1_1can__sniffer.html#a2342b933b01f118be030e93341908480',1,'canopen_sniffer.can_sniffer.stop()'],['../classcanopen__sniffer_1_1process__frame.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'canopen_sniffer.process_frame.stop()']]],
-  ['sync_78',['SYNC',['../classcanopen__sniffer_1_1frame__type.html#a03082caece11e4751c1e251f4a2525cd',1,'canopen_sniffer::frame_type']]]
+  ['save_5fframe_87',['save_frame',['../classcanopen__sniffer_1_1process__frame.html#a330d0aa4e889583805485fc069b8d7e3',1,'canopen_sniffer::process_frame']]],
+  ['save_5fframe_5fto_5fcsv_88',['save_frame_to_csv',['../classcanopen__sniffer_1_1can__sniffer.html#adea2a4942ab2e9b485f0f4acb670a172',1,'canopen_sniffer.can_sniffer.save_frame_to_csv()'],['../classcanopen__sniffer_1_1process__frame.html#a43c1277547b2e3a9db6631c8134c28f6',1,'canopen_sniffer.process_frame.save_frame_to_csv(self, int cob, frame_type ftype, int index, int sub, str name, str raw, str decoded)']]],
+  ['save_5fprocessed_5fframe_89',['save_processed_frame',['../classcanopen__sniffer_1_1process__frame.html#a8a398b1c620b611b45d8da94d9d35e6c',1,'canopen_sniffer::process_frame']]],
+  ['sdo_5fframes_90',['sdo_frames',['../classcanopen__sniffer_1_1DisplayCLI.html#a5b5c93be2f940dcfa114fa61f98c1d23',1,'canopen_sniffer::DisplayCLI']]],
+  ['sdo_5freq_91',['SDO_REQ',['../classcanopen__sniffer_1_1frame__type.html#aa227b2666e9ee4b630add48d6dec0c39',1,'canopen_sniffer::frame_type']]],
+  ['sdo_5fres_92',['SDO_RES',['../classcanopen__sniffer_1_1frame__type.html#a95373d5028d5cb3c1ce7b6120de91447',1,'canopen_sniffer::frame_type']]],
+  ['sdo_5fstats_93',['sdo_stats',['../classcanopen__sniffer_1_1bus__stats_1_1sdo__stats.html',1,'canopen_sniffer::bus_stats']]],
+  ['set_94',['set',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html#a112a711f766db446c2f517b794a1c04a',1,'canopen_sniffer::bus_stats::stats_data']]],
+  ['set_5fstart_5ftime_95',['set_start_time',['../classcanopen__sniffer_1_1bus__stats.html#a09ae5bd534d8dbec99a913ed75fe9848',1,'canopen_sniffer::bus_stats']]],
+  ['sparkline_96',['sparkline',['../classcanopen__sniffer_1_1DisplayCLI.html#a1716b97b8fa04faf79e5df60d9fb3489',1,'canopen_sniffer::DisplayCLI']]],
+  ['stats_97',['stats',['../classcanopen__sniffer_1_1process__frame.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer.process_frame.stats()'],['../classcanopen__sniffer_1_1DisplayCLI.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer.DisplayCLI.stats()']]],
+  ['stats_5fdata_98',['stats_data',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html',1,'canopen_sniffer::bus_stats']]],
+  ['stop_99',['stop',['../classcanopen__sniffer_1_1bus__stats.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'canopen_sniffer.bus_stats.stop()'],['../classcanopen__sniffer_1_1can__sniffer.html#a2342b933b01f118be030e93341908480',1,'canopen_sniffer.can_sniffer.stop()'],['../classcanopen__sniffer_1_1process__frame.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'canopen_sniffer.process_frame.stop()'],['../classcanopen__sniffer_1_1DisplayCLI.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'canopen_sniffer.DisplayCLI.stop()'],['../classcanopen__sniffer_1_1DisplayGUI.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'canopen_sniffer.DisplayGUI.stop()']]],
+  ['sync_100',['SYNC',['../classcanopen__sniffer_1_1frame__type.html#a03082caece11e4751c1e251f4a2525cd',1,'canopen_sniffer::frame_type']]]
 ];