+++ canopen_sniffer/dox/documentation/html/search/classes_5.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['sdo_5fstats_92',['sdo_stats',['../classcanopen__sniffer_1_1bus__stats_1_1sdo__stats.html',1,'canopen_sniffer::bus_stats']]],
-  ['stats_5fdata_93',['stats_data',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html',1,'canopen_sniffer::bus_stats']]]
+  ['payload_5fsize_115',['payload_size',['../classcanopen__sniffer_1_1bus__stats_1_1payload__size.html',1,'canopen_sniffer::bus_stats']]],
+  ['process_5fframe_116',['process_frame',['../classcanopen__sniffer_1_1process__frame.html',1,'canopen_sniffer']]]
 ];