+++ canopen_sniffer/dox/documentation/html/search/functions_e.js
@@ -1,5 +1,6 @@
 var searchData=
 [
-  ['update_5fsdo_5frequest_5ftime_123',['update_sdo_request_time',['../classcanopen__sniffer_1_1bus__stats.html#a2a6107d8e745cc7e16919053c332fea7',1,'canopen_sniffer::bus_stats']]],
-  ['update_5fsdo_5fresponse_5ftime_124',['update_sdo_response_time',['../classcanopen__sniffer_1_1bus__stats.html#a4d50a0e52bf7775a3c7b0906a4b8868a',1,'canopen_sniffer::bus_stats']]]
+  ['update_5frates_152',['update_rates',['../classcanopen__sniffer_1_1bus__stats.html#ab95b4bce7eb462bdaf41d216afbcfced',1,'canopen_sniffer::bus_stats']]],
+  ['update_5fsdo_5frequest_5ftime_153',['update_sdo_request_time',['../classcanopen__sniffer_1_1bus__stats.html#a2a6107d8e745cc7e16919053c332fea7',1,'canopen_sniffer::bus_stats']]],
+  ['update_5fsdo_5fresponse_5ftime_154',['update_sdo_response_time',['../classcanopen__sniffer_1_1bus__stats.html#a4d50a0e52bf7775a3c7b0906a4b8868a',1,'canopen_sniffer::bus_stats']]]
 ];