+++ canopen_sniffer/dox/documentation/html/search/variables_7.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['int_150',['int',['../classcanopen__sniffer_1_1bus__stats_1_1sdo__stats.html#a61569f2965b7a369eb10b6d75d410d11',1,'canopen_sniffer::bus_stats::sdo_stats']]],
-  ['interface_151',['interface',['../classcanopen__sniffer_1_1can__sniffer.html#ae6deea6e7d03b5d4db98ea209159e57f',1,'canopen_sniffer::can_sniffer']]]
+  ['hb_187',['HB',['../classcanopen__sniffer_1_1frame__type.html#a2e01ef077b0fedaea9e286ccdf2fe259',1,'canopen_sniffer::frame_type']]]
 ];