+++ canopen_sniffer/dox/documentation/html/search/variables_8.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['log_152',['log',['../classcanopen__sniffer_1_1bus__stats.html#ab09a63eb35b270b5cdbead1983ebdccb',1,'canopen_sniffer.bus_stats.log()'],['../classcanopen__sniffer_1_1eds__parser.html#ab09a63eb35b270b5cdbead1983ebdccb',1,'canopen_sniffer.eds_parser.log()'],['../classcanopen__sniffer_1_1can__sniffer.html#ab09a63eb35b270b5cdbead1983ebdccb',1,'canopen_sniffer.can_sniffer.log()'],['../classcanopen__sniffer_1_1process__frame.html#ab09a63eb35b270b5cdbead1983ebdccb',1,'canopen_sniffer.process_frame.log()'],['../namespacecanopen__sniffer.html#ab09a63eb35b270b5cdbead1983ebdccb',1,'canopen_sniffer.log()']]]
+  ['int_188',['int',['../classcanopen__sniffer_1_1bus__stats_1_1sdo__stats.html#a61569f2965b7a369eb10b6d75d410d11',1,'canopen_sniffer::bus_stats::sdo_stats']]],
+  ['interface_189',['interface',['../classcanopen__sniffer_1_1can__sniffer.html#ae6deea6e7d03b5d4db98ea209159e57f',1,'canopen_sniffer::can_sniffer']]]
 ];